+++ FindLoveByHobbys/src/main/java/com/fin/love/repository/profile/ProfileRepository.java
@@ -3,7 +3,7 @@ package com.fin.love.repository.profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileRepository extends JpaRepository<Profile, String> {
-	
-	
+
+		
 	
 }
