+++ FindLoveByHobbys/src/main/java/com/fin/love/service/ProfileService.java
@@ -1,0 +1,13 @@
+package com.fin.love.service;
+
+import org.springframework.stereotype.Service;
+
+import com.fin.love.repository.profile.ProfileRepository;
+
+@Service
+public class ProfileService {
+	
+	public ProfileRepository profileRepository;
+
+	
+}
