+++ FindLoveByHobbys/src/main/java/com/fin/love/web/ProfileController.java
@@ -3,17 +3,34 @@ package com.fin.love.web;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.fin.love.service.ProfileService;
+
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@RequiredArgsConstructor
 @Controller
+@RequestMapping("/profile")
 public class ProfileController {
 	
-	@GetMapping("/profile")
+	private ProfileService profileService;
+	
+	@GetMapping("/profiles")
 	public String home(Model model) {
 		log.info("home()");
 		
-		return "/profile/profile";
+		return "/profile/profiles";
 	}
+	
+	@GetMapping("/profileimage")
+	public String image(Model model) {
+		log.info("image()");
+		
+		return "/profile/profileimage";
+	}
+	
+	
 }
