+++ FindLoveByHobbys/src/main/resources/templates/profile/profileimage.html
@@ -9,16 +9,10 @@
 			<h1>프로필 이미지 등록 페이지</h1>
 			
 			<h1>프로필 이미지 등록</h1>
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
+				
 			
 			<h1>취미활동 이미지 등록</h1>
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
-				<img src="..." class="img-thumbnail" alt="...">
+
 		</div>
     </div> 
     </main>
